+++ source/gui/response_graph/DecibelMarkers.h
@@ -10,8 +10,8 @@ public:
     void paint(juce::Graphics& g) override;
 
 private:
-    float min_value_;
-    float max_value_;
+    int   min_value_;
+    int   max_value_;
     uint8 db_interval_;
 
     JUCE_DECLARE_NON_COPYABLE_WITH_LEAK_DETECTOR(DecibelMarkers)
